+++ models/todo.js
@@ -1,0 +1,21 @@
+const mongoose = require('mongoose');
+const {Schema} = mongoose;
+
+const todosScheme = new Schema({
+    text: {
+        type: String,
+        required: true
+    },
+    color: {
+        type: String,
+        default: 'red'
+    },
+    checked: {
+        type: Boolean,
+        default: false
+    }
+});
+
+const Todos = mongoose.model('Todos', todosScheme);
+
+module.exports = Todos;
